SHEP-575

+++ src/main/kotlin/com/bitclave/node/services/v1/OfferService.kt
@@ -121,16 +121,9 @@ class OfferService(
         strategy: RepositoryStrategyType
     ): CompletableFuture<List<Long>> {
         return supplyAsyncEx(Supplier {
-            val updatedOfferIds = offers.filter { it.id != 0L }.map { it.id }
-            val existedOffersByIds = offerRepository.changeStrategy(strategy).findByIds(updatedOfferIds)
-            val existedOffers = existedOffersByIds.map { it.id to it }.toMap()
             val readyForSaveOffers = offers.map {
-
-                val id = if (existedOffers.containsKey(it.id)) it.id else 0
-                val createdAt = if (id != 0L) existedOffers[id]?.createdAt ?: Date() else Date()
-
                 Offer(
-                    id,
+                    it.id,
                     owner,
                     it.offerPrices,
                     it.description,
@@ -140,10 +133,15 @@ class OfferService(
                     it.tags,
                     it.compare,
                     it.rules,
-                    createdAt
+                    it.createdAt
                 )
             }
-            val result = offerRepository.changeStrategy(strategy).saveAll(readyForSaveOffers)
+
+            var result: List<Offer> = listOf()
+            val saveAllTiming = measureTimeMillis {
+                result = offerRepository.changeStrategy(strategy).saveAll(readyForSaveOffers)
+            }
+//            logger.debug(" - save all timing $saveAllTiming")
 
             val prices = result.mapIndexed { index, offer ->
                 readyForSaveOffers[index].offerPrices.map { offerPrice ->
@@ -152,10 +150,19 @@ class OfferService(
                     price
                 }
             }.flatten()
-            offerPriceRepository.changeStrategy(strategy).saveAllPrices(prices)
 
-            val offersIdsForCleanupOfferSearches = readyForSaveOffers.filter { it.id != 0L }.map { it.id }
-            offerSearchService.deleteByOfferIds(offersIdsForCleanupOfferSearches, strategy)
+            val saveAllPricesTiming = measureTimeMillis {
+                offerPriceRepository.changeStrategy(strategy).saveAllPrices(prices)
+            }
+//            logger.debug(" - save all prices timing $saveAllPricesTiming")
+
+            val offersIdsForCleanupOfferSearches = offers.filter { it.id != 0L }.map { it.id }
+            val deleteOfferSearchTiming = measureTimeMillis {
+                if (offersIdsForCleanupOfferSearches.isNotEmpty()) {
+                    offerSearchService.deleteByOfferIds(offersIdsForCleanupOfferSearches, strategy)
+                }
+            }
+//            logger.debug(" - delete all offer searches by offerId timing $deleteOfferSearchTiming")
 
             result.map { it.id }
         })
